fix cards is not defined errors
