+++ js/js-structure/js-structure_quiz-app/components/CardList/CardList.js
@@ -1,0 +1,14 @@
+import Card from "../Card/Card.js";
+import { cards } from "../../utils/cards.js";
+
+export default function CardList() {
+  const cardList = document.createElement("section");
+  cardList.classList.add("card-list");
+
+  cards.forEach((card) => {
+    const cardElement = Card(card);
+    cardList.append(cardElement);
+  });
+
+  return cardList;
+}
