+++ js/js-structure/js-structure_quiz-app/index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Quiz-App</title>
     <link rel="stylesheet" href="./styles.css" />
-    <script src="./index.js" defer></script>
+    <script src="./index.js" type="module" defer></script>
   </head>
   <body>
     <div id="root"></div>
