+++ js/js-structure/js-structure_quiz-app/index.js
@@ -1,202 +1,7 @@
-// Store the cards in a global state
-const cards = [
-  {
-    question:
-      "In the Kingdom Hearts series who provides the english voice for Master Eraqus?",
-    answer: "Mark Hamill",
-    tags: ["kingdomheart", "mastereraqus", "english voices"],
-    isBookmarked: false,
-  },
-  {
-    question:
-      "Daniel Radcliffe became a global star in the film industry due to his performance in which film franchise?",
-    answer: "Harry Potter",
-    tags: ["hewhomustnotbenamed", "danielradcliffe", "abracadabra"],
-    isBookmarked: false,
-  },
-  {
-    question:
-      'In the game series "The Legend of Zelda" what is the name of the main character?',
-    answer: "Link",
-    tags: ["linktothepast", "linkbetweenworlds", "linkbetweenages"],
-    isBookmarked: false,
-  },
-];
-
-function Header() {
-  const header = document.createElement("header");
-  header.classList.add("header");
-  header.innerHTML = /* html */ `
-	  <h1 class="header__title">Quiz-App</h1>
-	`;
-
-  return header;
-}
-
-function handleFormSubmit(event) {
-  event.preventDefault();
-
-  const formData = new FormData(event.target);
-  const data = Object.fromEntries(formData);
-
-  const newCard = {
-    question: data.question,
-    answer: data.answer,
-    tags: data.tags.split(","),
-    isBookmarked: false,
-  };
-
-  cards.unshift(newCard);
-
-  // To prevent the need to rerender, we add the new card to the DOM directly
-  const cardList = document.querySelector(".card-list");
-  const newCardElement = Card(newCard);
-  cardList.prepend(newCardElement);
-
-  event.target.reset();
-  event.target.elements.question.focus();
-}
-
-function Form() {
-  const form = document.createElement("form");
-  form.classList.add("form");
-  form.innerHTML = /* html */ `
-		<label for="question">Your question:</label>
-		<textarea
-			id="question"
-			rows="7"
-			name="question"
-			maxlength="150"
-			data-js="input-question"
-		></textarea>
-		<span class="form__character-count">
-			<span data-js="amount-left-question"></span> characters left
-		</span>
-		<label for="answer">Your answer:</label>
-		<textarea
-			id="answer"
-			rows="7"
-			name="answer"
-			maxlength="150"
-			data-js="input-answer"
-		></textarea>
-		<span class="form__character-count">
-			<span data-js="amount-left-answer"></span> characters left
-		</span>
-		<label for="tags">Tags:</label>
-		<input id="tags" type="text" name="tags" />
-		<button class="form__submit-button" type="submit">Submit</button>
-	`;
-
-  const formFields = form.querySelectorAll('[data-js*="input"]');
-  const counterOutputs = form.querySelectorAll("[data-js*=amount-left]");
-
-  formFields.forEach((formField, index) => {
-    counterOutputs.forEach((output) => {
-      output.innerText = formField.maxLength;
-    });
-    formField.addEventListener("input", () => {
-      const currentAmountLeft = formField.maxLength - formField.value.length;
-      counterOutputs[index].innerText = currentAmountLeft;
-    });
-  });
-
-  form.addEventListener("submit", handleFormSubmit);
-
-  return form;
-}
-
-function Bookmark(props) {
-  const bookmark = document.createElement("button");
-  bookmark.classList.add("bookmark");
-
-  if (props.active) {
-    bookmark.classList.add("bookmark--active");
-  }
-
-  bookmark.type = "button";
-  bookmark.setAttribute("aria-label", "bookmark");
-  bookmark.innerHTML = /* html */ `
-		<svg
-			class="bookmark__icon"
-			xmlns="http://www.w3.org/2000/svg"
-			viewbox="0 0 24 24"
-		>
-			<path
-				d="M17 3H7c-1.1 0-2 .9-2 2v16l7-3 7 3V5c0-1.1-.9-2-2-2z"
-			/>
-		</svg>
-	`;
-
-  bookmark.addEventListener("click", props.onClick);
-
-  return bookmark;
-}
-
-function Card(props) {
-  const card = document.createElement("article");
-  card.classList.add("card");
-  card.innerHTML = /* html */ `
-		<h2 class="card__question" data-js="question"></h2>
-		<button class="card__button-answer" type="button" data-js="answer-button">
-			Show answer
-		</button>
-		<p class="card__answer" data-js="answer"></p>
-		<ul class="card__tag-list" data-js="tags"></ul>
-		<div class="card__button-bookmark" data-js="bookmark-button-wrapper"></div>
-	`;
-
-  const question = card.querySelector('[data-js="question"]');
-  const answerButton = card.querySelector('[data-js="answer-button"]');
-  const answer = card.querySelector('[data-js="answer"]');
-  const tags = card.querySelector('[data-js="tags"]');
-  const bookmarkButtonWrapper = card.querySelector(
-    '[data-js="bookmark-button-wrapper"]'
-  );
-
-  question.textContent = props.question;
-  answer.textContent = props.answer;
-
-  props.tags.forEach((tag) => {
-    const tagElement = document.createElement("li");
-    tagElement.classList.add("card__tag");
-    tagElement.textContent = "#" + tag;
-    tags.append(tagElement);
-  });
-
-  function handleBookmarkButtonClick(event) {
-    cards.find((card) => card.question === props.question).isBookmarked =
-      !props.isBookmarked;
-    // To prevent the need to rerender, we can just toggle the class here
-    event.currentTarget.classList.toggle("bookmark--active");
-  }
-
-  const bookmarkButton = Bookmark({
-    active: props.isBookmarked,
-    onClick: handleBookmarkButtonClick,
-  });
-  bookmarkButtonWrapper.append(bookmarkButton);
-
-  function handleAnswerButtonClick() {
-    answer.classList.toggle("card__answer--active");
-  }
-
-  answerButton.addEventListener("click", handleAnswerButtonClick);
-
-  return card;
-}
-
-function CardList() {
-  const cardList = document.createElement("section");
-  cardList.classList.add("card-list");
-
-  cards.forEach((card) => {
-    const cardElement = Card(card);
-    cardList.append(cardElement);
-  });
-
-  return cardList;
-}
+import { cards } from "./utils/cards.js";
+import Header from "./components/Header/Header.js";
+import Form from "./components/Form/Form.js";
+import CardList from "./components/CardList/CardList.js";
 
 function App() {
   const app = document.createElement("main");
