+++ js/js-structure/js-structure_quiz-app/utils/cards.js
@@ -1,0 +1,24 @@
+// Store the cards in a global state
+export const cards = [
+  {
+    question:
+      "In the Kingdom Hearts series who provides the english voice for Master Eraqus?",
+    answer: "Mark Hamill",
+    tags: ["kingdomheart", "mastereraqus", "english voices"],
+    isBookmarked: false,
+  },
+  {
+    question:
+      "Daniel Radcliffe became a global star in the film industry due to his performance in which film franchise?",
+    answer: "Harry Potter",
+    tags: ["hewhomustnotbenamed", "danielradcliffe", "abracadabra"],
+    isBookmarked: false,
+  },
+  {
+    question:
+      'In the game series "The Legend of Zelda" what is the name of the main character?',
+    answer: "Link",
+    tags: ["linktothepast", "linkbetweenworlds", "linkbetweenages"],
+    isBookmarked: false,
+  },
+];
